client-api/requirements.txt
fastapi[all]
numpy
azure-storage-queue
azure-cosmos
azure-storage-blob
